Add PathsWithModules helper function
PathsWithModules returns any directories containing files with a `.yang`
extension under the root directory passed. Added to facilitate the
import of a directory tree of YANG modules.

Add a new YANG test module and test to confirm directories are recursed
into and that paths added this way can be used to import modules within
them.