entry: Fix parents when merging a usage
By setting the parent on the entries as they merge, the correct paths
are shown. Previously, the name of the grouping was also seen, whereas
this is not part of the path as described by XPath or the JSONPath specs.

Add a regression test to confirm we see the correct path with this
change to entry.go, without the name of the grouping.