Add the Extra map to an Entry.
The Extra map is a map of all fields that have no specific processing.
The map is keyed on the field name (such as "revision").  The value
is a slice of interface{} (in case the statement is seen more than
once).  It is the responsibility of client code to map the elements
of the slice to their approrpiate type (e.g., *yang.Revision).