yang: Use portable path functions from filepath
The "path" package's Base, Dir and Join functions use the Unix path
separator '/', which is not compatible with Windows. Use instead the
same named functions from the "path/filepth" package, which use the
system's path separator.

This change causes the test suite to pass on Windows 10 and should
similarly affect earlier versions of Windows.