yang: Add support for YANG 1.1 "anydata" statement
Support reading the YANG 1.1 "anydata" statement.

The "anydata" statement is defined at:
https://tools.ietf.org/html/rfc7950#section-7.10

Prior to this change, the presence of an "anydata"
sub-statement in YANG modules leads to errors such as:

  ietf-restconf.yang:196:9: unknown list field: anydata
  ietf-yang-patch.yang:263:9: unknown list field: anydata

The "anydata" statement is a YANG data node statement, similar
in nature to the YANG 1.0 "anyxml" statement. It stores an
opaque value whose use and encoding is agreed upon apriori
between parties using the field.

Add lexing tests for the "anydata" and "anyxml" statements.