yang: Add support for YANG 1.1 "action" statement
Support YANG 1.1's "action" statement in goyang.

The "action" statement is defined at:
https://tools.ietf.org/html/rfc7950#section-7.10

Prior to this change, the presence of an "action"
sub-statement in YANG modules leads to errors such as:

  ietf-routing@2016-11-04.yang:447:9: unknown list field: action
  ietf-keychain.yang:129:9: unknown list field: action

The "action" statement has the same sub-statements as the "rpc"
statement. The difference is that instead of defining an RPC
action found only at the module level, "action" can appear as
sub-statements to "container" and "list" data nodes, as well as
"grouping" and "augment" statements similarly.

This change does the following:
  - Define the Action type (in yang.go), as an RPC.
  - Add Action as a 0..n cardinality child field to:
    Container, List, Grouping and Augment types.
  - Add tests for successful and unsuccessful module lexing and
    ToEntry parsing of RPC and Action types.