panic if the lexer items channel overflows
To avoid a deadlock when sending a *token to the lexer's items
channel - which can occur if too many errors occur within a single
token - instead panic when the items channel fills up.

Non-error items (i.e., normal tokens) should not overflow the channel,
so use a different panic message (indicating a library bug) if this
occurs.