Parse and semantic-check min-elements and max-elements.
Currently, the statements don't undergo semantic checking, forcing
backends like ygot to have to handle it (which incidentally does it
wrong -- to be handled separately).

These fields are the most useful as uint64. This is because whether or
not they're explicitly set in the `list` or `leaf-list`, there is always
a valid uint64 value for them. As a result, uint64 is proposed as the
type for these fields.

By not using a pointer, however, the implementation of deviations
necessitates the new `deviationPresence` field to distiguish, for a
DeviationEntry, between no deviation applied vs. the default uint64
value. So, it has been added to the `Entry` type.

Lastly, deviation for deletes was actually unimplemented due to a test
bug that misinterpreted `nil` as "don't test it", rather than "test that
it's nil". This has been implemented and the test fixed.