Fix lexer to identify single-quote string that immediately occurs after a concatenation (plus sign).
This code makes the new assumption that a `'` that follows `+` must be a string.

This change will break YANG files where there exists an unquoted string that contains a `+'` character substring. Given that this is unlikely, I think this is a fine tradeoff to get cases like #180 working.

Indeed, the current lexer is already assuming `"` cannot exist within an unquoted string (see `lexIdentifier`). This is already a much more restrictive assumption.

NOTE: In YANG1.1:
```
An unquoted string cannot contain any single or double quote
      characters.
```

Fixes #180.