Export Module.modules
Downstream code that tests using Entry objects (e.g. ygot) might
be blocked by this if they are manually created, since `modules` can
only be populated when Entry objects are created using the `Modules`
object and methods.

I don't see a reason for keeping it unexported, so creating this PR.