Ignore extra error tokens.
An alternative that fixes #116.

This is essentially what @pborman proposed in
https://github.com/openconfig/goyang/pull/117#issuecomment-616119988.

I wasn't able to reproduce test failing as mentioned by @andaru in
https://github.com/openconfig/goyang/pull/117#issuecomment-622232317 in
the current codebase, so pborman's suggestion now makes more sense.

The reason seems to be that error tokens are always skipped, so they
don't seem to affect `Location` as Andrew mentioned:
https://github.com/openconfig/goyang/blob/1fb7893798e04830b4d7fbbcd6672b95668fcf73/pkg/yang/parse.go#L207-L214