feat(yangentry): implement ParseWithOptions method
This code is a Contribution to the OpenConfig goyang sub-project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind.

Signed-off-by: Vilmos Martinek <vilmos.martinek@nokia.com>